+++ sample/src/main/java/io/snabble/testapp/HomeFragment.java
@@ -69,6 +69,13 @@ public class HomeFragment extends Fragment {
             }
         });
 
+        v.findViewById(R.id.age_verification).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                ((BaseActivity)getActivity()).showAgeVerification();
+            }
+        });
+
         v.findViewById(R.id.clear_cache).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {